requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
PyYAML>=6.0.2
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0
langchain-milvus>=0.1.8
pymilvus>=2.4.5
pypdf>=5.0.0
unstructured[pdf]>=0.16.0
prometheus-fastapi-instrumentator>=7.0.0
rank-bm25>=0.2.2
tenacity>=9.0.0
pytest>=8.2.0
